update tests to support freebsd
